Miscellaneous Setup
Played around a lot with several things, some ideas not complete yet,
like collision. Tried to set up how things will be executed once more of
the core system is developed. Sam I'm planning on showing up Tuesday
with a bit of a laundry list of stuff for us to plan out with
organization and such. Currently in this build I have a player and enemy
rendered that take turns moving tile by tile, a setup for handling
ability usage, and I added floor and wall classes which currently only
extend tile and set canCollide to true or false.