Full melee combat, Facing, Round timed buffs.
Alright, we have completely functional melee combat, q and w are attacks
and e is a special move. Keep in mind it might feel a bit wonky without
collision or dynamic sprites since it relies on facing, but just know
you face in the direction you last moved. I plan to hopefully get some
knid of ranged combat system done tomorrow. Some things to note:

I probably could have done collision, but it's 6:00 and I want to relax
for the rest of my evening so I'll just do it tomorrow.

Don't try walking out of screen bounds because it causes a crash, I
didn't bother making a fix for that since we'll probably have walls
there in the final product.

I used a metric shitton of print statements for debugging but most of
them are commented out. Point is I still left them there for anyone who
wants to see a bit more of what's happening.

As usual, let me know if you need anything else, I'll be lurking on
github for the rest of the night, though I may not notice posts right
away.