Added Item Functionality
Added basic collision detection with items, inventory, and item effects.

Note: Items, at least ones with passive effects, are pretty damn easy to
make. I made some simple stat-boosting ones, but all you need is a name,
an effect, and the inverse of that effect to make a cool magic item. I
put them all inside of Item itself, because literally nothing else is
going to access that.

NOTE: We still need a way to drop an item from the player. I built a
method for it, but that method is vestigial, should be in UI somewhere.