Added animation, cleaned up.
I added the appropriate code for animation of movement, but it doesn't
actually do anything yet. Currently with movement, the Player just moves
a tile and there is no room in between for animation.. Also, I cleaned
up my Assets class a bit, and used the auto-organize imports hotkey on
the entire project.