Combat Log, Stair Code in Place, Still Need to Draw.
Alright, I have the combat log working, David can I ask for your help in
drawing the stairs? All they are is just a tile with a boolean isStairs
set to true. I use a method in Map called makeStairs to make them, which
determines the positon of the tile that is set to be the stairs. If I
could just as you to draw a square on top of that spot, that would be
awesome, and I could then complete the code for going down floors.